worker-service/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
face-recognition==1.3.0
numpy==1.24.3
boto3==1.34.10
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
